pyproject.toml
[project]
name = "agent-capstone"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk[http-server,mysql,sqlite]>=0.3.19",
    "fastmcp>=2.13.1",
    "google-adk>=1.19.0",
    "langchain>=1.1.0",
    "mcp[cli,server]>=1.22.0",
    "pandas>=2.3.3",
    "pypdf>=6.4.0",
]
